Separated config_dialog.c
